# Removed Memory._id