--- a/tests/Benchmarks/IndexOf.cs
+++ b/tests/Benchmarks/IndexOf.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Xunit;
 using Microsoft.Xunit.Performance;
 using System;
 using System.Buffers;
@@ -10,18 +11,22 @@ using System.Text;
 
 public class IndexOfBench
 {
-    static int s_bufferLength = 1000;
+    static int s_bufferLength = 2000;
     static byte[] s_buffer = new byte[s_bufferLength];
     static int s_loops = 1000;
 
-    static IndexOfBench() 
-    {
-        s_buffer[s_bufferLength - 100] = 255;
-    }
-
     [Benchmark]
-    static int SpanIndexOf()
+    [InlineData(0)]
+    [InlineData(1)]
+    [InlineData(2)]
+    [InlineData(4)]
+    [InlineData(8)]
+    [InlineData(16)]
+    [InlineData(30)]
+    [InlineData(1000)]
+    static int SpanIndexOf(int at)
     {
+        s_buffer[at] = 255;
         Span<byte> buffer = s_buffer;
         int index = 0;
         foreach (var iteration in Benchmark.Iterations)
@@ -33,13 +38,23 @@ public class IndexOfBench
                 }
             }
         }
+        s_buffer[at] = 0;
         return index;
     }
 
     [Benchmark]
-    static int VectorizedIndexOf()
+    [InlineData(0)]
+    [InlineData(1)]
+    [InlineData(2)]
+    [InlineData(4)]
+    [InlineData(8)]
+    [InlineData(16)]
+    [InlineData(30)]
+    [InlineData(1000)]
+    static int VectorizedIndexOf(int at)
     {
         if(!Vector.IsHardwareAccelerated) return 0;
+        s_buffer[at] = 255;
 
         Span<byte> buffer = s_buffer;
         int index = 0;
@@ -52,6 +67,7 @@ public class IndexOfBench
                 }
             }
         }
+        s_buffer[at] = 0;
         return index;
     }
 }