@ahsonkhan 2017-02-23
Remove the VerifyId method?
https://github.com/dotnet/corefxlab/blob/master/src/System.Slices/System/Buffers/OwnedMemory.cs#L213