@davidfowl 2017-02-23
You need to remove a bunch of tests for use after free.