@joshfree 2017-02-23
Would it make sense instead of removing _id completely to put it under #if DEBUG for now?