@KrzysztofCwalina 2017-02-23
@joshfree, I was trying to, but it's all over signatures. The source would get very hard to read. I can add it back later in a couple of hours.